[#503] Update Cardano DB Sync version to 4.1.0
In the context of the user story, which aimed to update the Cardano DB
Sync version for Sanchonet to unblock tasks related to verification and
adjustments, this commit bumps the version of Cardano DB Sync to 4.1.0
on a development stack providing only Cardano tools.

By upgrading to the latest release, developers will be able to proceed
with the necessary verifications and adjustments to ensure application
compatibility with the updated DB Sync version.

Changes:
- Updated the `docker-compose.node+dbsync.yml` file to use the Cardano
  DB Sync image version 4.1.0
  (`ghcr.io/intersectmbo/cardano-db-sync:sancho-4.1.0`).