[#526] Add modules .envrc files
The purpose of this commit is to create .envrc files within specific
modules that reference the root .envrc file and utilize the
corresponding flake for configuration. By establishing these .envrc
files in the `backend`, `infra`, and `scripts` modules, the project
enhances the organization and management of application modules using
Nix Flakes. This setup ensures that each module aligns with the root
configuration, streamlining the usage of Nix Flakes for combining
application modules effectively within the project structure.

Changes:
- Renamed .env.example to .envrc.example, ensuring consistency in naming
  conventions.
- Added a new .envrc file in the `govtool/backend` directory,
  referencing the root environment and using the corresponding flake for
  the backend module.
- Created a .envrc file in the `infra/terraform` directory, linking to
  the root environment and utilizing the appropriate flake for the
  infrastructure module.
- Introduced a .envrc file in the `scripts/govtool` directory, pointing
  to the root environment and leveraging the matching flake for the
  scripts module setup.