[#530] Add stylish-haskell to the developer's toolset
In line with the user story's objective of implementing code formatting
and checks for the backend to ensure a consistent code style, this
commit focuses on enhancing the developer's configuration in the backend
project. By adding stylish-haskell to the tools dependencies, backend
developers will now have access to this code formatting tool, enabling
them to maintain a standardized code style effortlessly.

The modification in this commit involves updating the default.nix file
within the govtool/backend directory. The addition of stylish-haskell to
the list of tools dependencies ensures that developers can utilize this
tool to enforce consistent code formatting standards, thereby
contributing to a unified and coherent codebase across the backend
project.