[#530] Add hlint to the developer's toolset
In alignment with the user story's objective to implement code
formatting and checks for the backend to ensure a consistent code style,
this commit includes the addition of hlint to the project's build tools
within the default.nix file. By introducing hlint to the backend's
development environment, developers can leverage this static code
analysis tool to further enhance the code quality and maintain adherence
to the defined code style standards.

The main change in this commit involves editing the default.nix file
located in the govtool/backend directory. Specifically, the modification
adds hlint to the list of build tools, enabling backend developers to
utilize this tool for static code analysis alongside other existing
tools like cabal-install, haskell-language-server, and stylish-haskell.
This enhancement contributes to the establishment of a comprehensive
code formatting and checking mechanism, promoting code consistency and
quality within the backend project.