[#555] Update Cardano Node to version 8.9.0
The Cardano Node version has been updated to 8.9.0 to align with the latest configuration requirements specified on https://book.world.dev.cardano.org. This update was necessary as the existing configuration was found to be incompatible with previous version 8.8.0-pre, requiring an intermediate update to ensure the proper functioning of the Cardano Node.