[#624] Use newer version of checkout action
The user story described a bug in the deployment setup preventing the
continuous deployment (CD) of the metadata-validation service. The issue
stemmed from the deployment environment lacking access to the directory
housing the necessary Dockerfile and its context. To resolve this, the
deployment setup needed adjustments to enable the CD process for the
metadata-validation service, ensuring the latest changes are reflected
in the production environment.

In this commit, the version of the checkout action in the
build-and-deploy-dev.yml workflow file was updated from v3 to v4. This
modification aims to address the bug by utilizing the newer version of
the checkout action, potentially resolving the access issues to the
directory containing the Dockerfile and its context, hence facilitating
the successful building and deployment of the metadata-validation
service in the CD pipeline.