[#434] Ensure deployment triggered only by specific tag
The commit adjusts the `.github/workflows/build-and-deploy-test.yml`
file to restrict the triggering of test deployments solely by specific
tags within the repository. The modifications include altering the
workflow to activate upon tag creation events that match the specified
criteria, safeguarding that deployments to the test server are solely
initiated by tag creation events. This change reinforces adherence to
the new Git workflow proposal, guaranteeing that deployments to the test
environment occur efficiently and accurately by aligning with the
revised deployment process.