[#715] Update Traefik image version to v3.0
The commit modifies the Traefik image version specified in the
docker-compose.yml.tpl file from v2.10 to v3.0. By updating the Traefik
image, we ensure compatibility with the necessary features and
improvements introduced in version 3.0. This adjustment aligns with the
user story's goal to enhance load balancing and routing setup in the dev
environment by making sure that the configuration utilizes the latest
Traefik version for more efficient backend service management and
testing capabilities.