[#715] Add additional CORS exception for dev environment
The changes made in this commit focus on enhancing the CORS
configuration for the dev environment by adding an additional exception
to allow specific origins. The modification includes appending
"http://localhost:5173" to the list of allowed hosts when the
environment is set to "dev." This adjustment aims to address
Cross-Origin Resource Sharing (CORS) issues related to accessing
services from different origins, ensuring that the specified host is
permitted to interact with the backend service seamlessly.