[#714] Update Cardano DB Sync with fixed configuration files
The commit updates the Cardano DB Sync component in the Makefile by
changing the image tag from "sancho-4-2-0" to "sancho-4-2-1." This
adjustment ensures that the DB Sync is now set to the correct version,
addressing any previously encountered issues and guaranteeing the proper
functioning of the network on SanchoNet.

This was necesary since there was an issue with the previous build with
non parsing configuration.