[#742] Create a status-service AWS registry
The user story described a situation where the deployment to the Dev
environment was failing due to missing context for the status-service.
The expected behavior was for the deploy to succeed. In response, a
status-service AWS registry was created to address this missing context,
ensuring that the deploy to Dev can now successfully utilize the
status-service.