[#742] Create a metadata-validation AWS registry
The user story indicated that the deployment to the dev environment was
failing due to a missing context for the status-service. To address this
issue and enable a successful deployment, a status-service AWS registry
was created. However, to further enhance the deployment process, a
metadata-validation registry has been implemented. This addition ensures
that the necessary context is available for the deployment, optimizing
the functionality and reliability of the Dev environment.