[#742] Add status service docker image utilization to the deploy stack
The user story highlighted the issue of a failed deployment to the dev
environment due to a missing context for the status-service. The
expected behavior was for the deploy to succeed. To address this,
changes were made to incorporate the status-service docker image
utilization into the deploy stack. The addition of the status-service
AWS registry now ensures that the necessary context is available for the
deployment process, moving the build process to the machine where the
deploy is performed, allowing for a successful deployment to the dev
environment.