[#742] Add metadata-validation docker image utilization to the deploy stack
The user story highlighted the issue of a failed deployment to the dev
environment due to a missing context for the status-service. The
expected behavior was for the deploy to succeed. In response to this,
changes have been made to integrate the metadata-validation docker image
into the deploy stack. By creating the necessary context for the
metadata-validation registry, the deployment process can now utilize
this image efficiently. The modifications include new files like
`Makefile` in `govtool/metadata-validation`, adjustments in
`scripts/govtool/Makefile` to incorporate the metadata-validation
registry, and changes in `docker-compose.yml.tpl` and `info.mk` to
include metadata-validation details for enhanced deployment
functionality.