[#849] Add a registry for the analytics-dashboard module
A registry has been incorporated for the analytics-dashboard module
within the Terraform configuration. This modification introduces a new
module named "govtool-ecr-analytics-dashboard" sourced from
"./modules/ecr" with the repository name set to "analytics-dashboard."
By including this new module, the analytics-dashboard component is now
integrated into the infrastructure setup, aligning with the user story's
objective of configuring the sanchogov.tools sub-domain for the
analytics service.