[#849] Add a secret to analytics-dashboard
Introduce a secret configuration for the analytics-dashboard service
within the Docker Compose setup. This modification includes the addition
of the `google-credentials.json` secret to enhance the security and
integrity of sensitive information utilized by the service. By
incorporating the secret, the analytics-dashboard service gains access
to the specified Google credentials, ensuring secure interactions with
external services and APIs, as mandated by the user story's requirement
to configure the `sanchogov.tools` sub-domain for the analytics service.