[#849] Fix broken docker compose
In this commit, the docker-compose file was modified to include the
`google-credentials.json` file as a secret volume. This change was made
to ensure that the necessary Google credentials are available for the
analytics service running on the `participation.sanchogov.tools`
sub-domain. Including the `google-credentials.json` file in the
docker-compose configuration is crucial for the proper functioning of
the analytics dashboard service.