[#849] Add build arg  for docker image for analytics-dashboard
In this commit, an argument (`NEXT_PUBLIC_API_URL`) was added to the
`Dockerfile` of the analytics-dasboard service. This change is necessary
to dynamically specify the API URL that the dashboard should connect to.
By introducing this build arg, the analytics dashboard service can now
adapt to different environments, like the
`participation.sanchogov.tools` sub-domain, as part of configuring the
analytics service under one domain tree.