Fix building analytics image for multiple environments
The current commit modifies the script `analytics-dashboard.mk` in the
`govtool` directory to address the issue of building the analytics image
for multiple environments. Previously, the image tag was being generated
based on the latest commit hash for the `analytics-dashboard`, which was
not sufficient for distinguishing between different environments. To
solve this, the script now appends the `$(env)` variable to the image
tag, allowing for unique identification of images across various
environments.