[#262] Add an overview of the infrastructure in the documentation
The commit adds a new file `infra/terraform/README.md` providing an
overview of the Terraform infrastructure for the GovTool project. The
documentation includes details on the architecture, components like IAM
Groups, DNS management, ECR repositories, and EC2 instances, as well as
configurations, managed resources, prerequisites for managing the
infrastructure, and instructions for initialization and modification of
components. It also outlines the CI/CD support, outputs, and guidelines
for contributions.