[#1195] Remove unnecessary deployment config dependancy
In the provided user story, the goal was to troubleshoot the failure in
the Toggle Maintenance Page workflow by ensuring that all necessary
environment variables are present. The acceptance criteria focused on
identifying and setting the required environment variables for the
workflow to run successfully.

However, the commit candidate addresses this by removing the unnecessary
creation and upload of the deployment config for the toggle-maintenance
target. By simplifying the process and adjusting the workflow
configuration, the changes aim to streamline the execution of the toggle
maintenance page workflow and prevent potential errors caused by missing
environment variables that are not really required.