[#1538] Update Cardano Node and DB Sync versions in dev environment
Bumped the versions of Cardano Node and Cardano DB Sync in the
development environment to ensure compatibility with the latest features
and fixes. Updated the `docker-compose.node+dbsync.yml` file to use
`cardano-node:9.0.0-sancho` and `cardano-db-sync:sancho-5.1.0`. This
change addresses the need to stay current with upstream improvements and
ensures that our development environment remains in line with the latest
network protocols.