EnableFutureGenesis option added to dbsync config
Updated the dbsync configuration to include the "EnableFutureGenesis"
option. Modified the `scripts/govtool/config.mk` file to add
`db-sync-config.json` under the `cardano_configs` list. Also included a
new command to download the `db-sync-config.json` file and append the
"EnableFutureGenesis" configuration to it. This change allows the
synchronization process to support future genesis blocks, making the
system more adaptable to upcoming updates in the Cardano blockchain.