Add missing ENV variable for PDF API in workflow
Updated the GitHub Actions workflow to include the `PDF_API_URL`
environment variable. This ensures that the resync process has access to
the necessary endpoint for generating PDFs, improving the overall
functionality and integration of the resync process within the Cardano
node and DBSync synchronization workflow. This change addresses the
missing configuration and enhances the reliability of the workflows.

Additionally, this commit introduces the recent adjustments made to the
workflow that deploys dev environment.