Bump Cardano Node version to 9.2.1 on sanchonet
Updated the Cardano Node version from 9.1.1 to 9.2.1 in both the
`Makefile` and `docker-compose.node+dbsync.yml` files to ensure
compatibility with the latest release. This change is in line with the
user story requirement to use version 9.2.1 on sanchonet in the proposal
pillar, addressing potential improvements or fixes that come with the
new version.